requirements.txt
rich>=13.7.0
typer>=0.9.0
jinja2>=3.1.2
pyyaml>=6.0
requests>=2.31.0
dnspython>=2.4.0
redis>=5.0.0
psutil>=5.9.0
python-systemd>=235